refactor: 패stpe3, step4 패키지 í통합, 포맷팅 ìë³적용
