+++ src/main/java/racing/Cars.java
@@ -1,4 +1,4 @@
-package step4;
+package racing;
 
 import java.util.ArrayList;
 import java.util.List;
