+++ src/main/java/racing/InputView.java
@@ -1,6 +1,5 @@
-package step4;
+package racing;
 
-import java.util.List;
 import java.util.Scanner;
 
 public class InputView {
