--- src/main/java/step4/Dice.java
@@ -1,6 +1,0 @@
-package step4;
-
-public interface Dice {
-
-    int makeNumber();
-}
