--- src/main/java/step4/DiceWithInput.java
@@ -1,16 +1,0 @@
-package step4;
-
-public class DiceWithInput implements Dice {
-
-    private int number;
-
-    public DiceWithInput(int input){
-        this.number = input;
-    }
-
-    @Override
-    public int makeNumber() {
-
-        return this.number;
-    }
-}
