--- src/main/java/step4/DiceWithRandom.java
@@ -1,14 +1,0 @@
-package step4;
-import java.util.Random;
-
-public class DiceWithRandom implements Dice {
-
-    private final static int MAX_NUMBER = 10;
-    Random random = new Random();
-
-    @Override
-    public int makeNumber() {
-
-        return random.nextInt(MAX_NUMBER);
-    }
-}
