+++ src/test/java/racing/ParserUtilsTest.java
@@ -1,11 +1,9 @@
-package step4;
+package racing;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.junit.jupiter.api.Assertions.*;
 
 class ParserUtilsTest {
